use sample counter from driver, code cleanup
The sample counter from the driver is now used for
timestamp estimation, which should give us a better
timestamp estimate in case of lost samples.

Also a lot of unneeded code (as now provided by the
TimestampEstimatorStatus) for statistics was removed.

Finally the error state of timestamp mismatch was removed
as there is no real way of knowing if they have really diverged
with the current codebase. Or to say it the other way around,
the check triggered falsely in the burst case.